rag/requirements.txt
beautifulsoup4==4.13.4
ipython==8.12.3
langchain==0.3.25
langchain_chroma==0.2.4
langchain_community==0.3.24
langchain_core==0.3.60
langchain_ollama==0.3.3
langchain_tavily==0.1.6
langchain_text_splitters==0.3.8
langgraph==0.4.5
python-dotenv==1.1.0
typing_extensions==4.13.2
